% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synsyn.R
\name{reduce_motion}
\alias{reduce_motion}
\title{Summarize individual motion data iteratively}
\usage{
reduce_motion(bd_motion, fn, reduce_fn, head_n = Inf)
}
\arguments{
\item{bd_motion}{A BIDS motion files table, of the format created by `bids::bids_motion`}

\item{fn}{A function that takes arguments `f(file_path, session_id, participant_id, task_label)` and returns a data frame}

\item{reduce_fn}{A function that takes arguments `f(result, step_result)`, the first is the accumulated value, and the second is the result of `fn` for the current element in `bids_motion`.}

\item{head_n}{How many pairs to compute before returning}
}
\description{
Summarize individuals' motion data into a single frame with a custom function using the functional programming approach called reduce, fold, etc.
}
