% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synsyn.R
\name{summarize_motion}
\alias{summarize_motion}
\title{Summarize individuals' motion data}
\usage{
summarize_motion(bd_motion, fn, head_n = Inf)
}
\arguments{
\item{bd_motion}{A BIDS motion files table, of the format created by `bids::bids_motion`}

\item{fn}{A function with arguments `f(file_path, session_id, participant_id, task_label)` that returns a summarized data frame}

\item{head_n}{How many pairs to compute before returning}
}
\description{
Given a bids dataset, apply a function to each individual's sessions of motion, one at a time.
}
