% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synsyn.R
\name{summarize_motion_pairs}
\alias{summarize_motion_pairs}
\title{Summarize pairs of motion data}
\usage{
summarize_motion_pairs(
  bd_motion,
  relate,
  preprocess,
  group_summary = bind_rows,
  ordered = F,
  unique = NA,
  self_relate = F,
  head_n = Inf,
  progress = T
)
}
\arguments{
\item{bd_motion}{A BIDS motion files table, of the format created by `bids::bids_motion`}

\item{relate}{A function with two arguments that relate two participant's motion together into a data frame of values.}

\item{preprocess}{A function with arguments `f(file_path, session_id, participant_id, task_label)` that returns a data frame on which `relate` operates. At the very least, this function needs to read the file at `file_path` into a data frame.}

\item{ordered}{Whether the pairs are ordered pairs, i.e, whether to compute both `relate(A, B)` and `relate(B, A)` for participants A and B.}

\item{unique}{Deprecated, use `self_relate` instead}

\item{self_relate}{Whether to compute relationships within the same person, i.e., whether to calculate `relate(A, A)` for a participant A.}

\item{head_n}{How many pairs to compute before returning.}

\item{progress}{Whether to show a progress bar}
}
\description{
Given a bids dataset, apply a function relating pairs of participants in the same session.
}
